feat(header): create main nav component
