+++ src/components/layout.js
@@ -9,20 +9,11 @@ import './layout.css'
 import { useStyles } from '../utils'
 
 const Layout = ({ children }) => {
-  const data = useStaticQuery(graphql`
-    query SiteTitleQuery {
-      site {
-        siteMetadata {
-          title
-        }
-      }
-    }
-  `)
   const { layout } = useStyles()
 
   return (
     <>
-      <Header siteTitle={data.site.siteMetadata.title} />
+      <Header />
       <Grid container direction="row" justify="space-between" className={layout}>
         <Grid item xs={12}>
           <main>{children}</main>
